+++ lib/flutter_advanced_canvas_editor.dart
@@ -3,6 +3,7 @@ import 'flutter_advanced_canvas_editor_platform_interface.dart';
 
 export 'src/canvas_controller.dart';
 export 'src/canvas.dart';
+export 'src/canvas_layer.dart';
 
 class FlutterAdvancedCanvasEditor {
   Future<String?> getPlatformVersion() {